+++ src/View/ActionPanelWidget.h
@@ -17,11 +17,13 @@ public:
     void updateForPhase(PhaseType phase, bool hasPendingAction);
     void updateForPendingAction(const PendingActionData& info);
     void setHint(const QString& text);
+    void setSkillAvailable(bool available);
 
 signals:
     void playPhaseEnded();
     void respondSkipped();
     void discardConfirmed();
+    void skillRequested();
 
 private:
     void setupUi();
@@ -31,6 +33,7 @@ private:
     QPushButton* m_endPlayBtn;
     QPushButton* m_skipResponseBtn;
     QPushButton* m_confirmDiscardBtn;
+    QPushButton* m_skillBtn;
 };
 
 #endif // ACTIONPANELWIDGET_H